move sibyl-to-stream to separate file
