+++ package.json
@@ -37,7 +37,8 @@
     "send": "^0.15.2",
     "sheetify": "^6.1.0",
     "stencila": "0.27-preview",
-    "tachyons": "^4.7.4"
+    "tachyons": "^4.7.4",
+    "validate-formdata": "^1.0.0"
   },
   "devDependencies": {
     "minidocs": "^4.2.3",
