+++ server/server.js
@@ -1,5 +1,4 @@
 const cookie = require('cookie')
-const eos = require('end-of-stream')
 const jwt = require('jsonwebtoken')
 const merry = require('merry')
 const path = require('path')
@@ -8,6 +7,7 @@ const send = require('send')
 const spawn = require('child_process').spawn
 const url = require('url')
 
+const sibylToStream = require('./sibyl-to-stream')
 const errors = require('./errors')
 
 const env = {
@@ -92,66 +92,6 @@ app.route('default', errors.EURLNOTFOUND)
 
 // Start the app
 app.listen()
-
-// Safely forward the sibyl script into a stream of messages
-// Connects the sibyl child process, a write stream and a context
-function sibylToStream (sibyl, req, res, ctx) {
-  var closed = false
-
-  sibyl.stdout.on('data', onStdout)
-  sibyl.stderr.on('data', onStderr)
-  sibyl.on('exit', onExit)
-
-  eos(res, function (err) {
-    if (err) ctx.log.error(err)
-    ctx.log.debug('closing SSE stream')
-    sibyl.stdout.removeListener('data', onStdout)
-    sibyl.stderr.removeListener('data', onStderr)
-    sibyl.removeListener('exit', onExit)
-    closed = true
-  })
-
-  function onStdout (data) {
-    if (closed) return
-    for (let line of data.toString().split('\n')) {
-      if (line.length) {
-        const match = line.match(/^(STEP|IMAGE|GOTO) (.+)$/)
-        if (match) {
-          let type = match[1]
-          let data = match[2]
-          if (type === 'STEP') {
-            ctx.log.debug('SSE: sending step')
-            res.write(`event: step\ndata: ${data}\n\n`)
-          } else if (type === 'IMAGE') {
-            ctx.log.debug('SSE: sending image')
-            res.write(`event: image\ndata: ${data}\n\n`)
-          } else if (type === 'GOTO') {
-            const token = jwt.sign({ url: data }, ctx.env.TOKEN_SECRET, { expiresIn: '12h' })
-            ctx.log.debug('SSE: sending stdout goto')
-            res.write(`event: goto\ndata: ${token}\n\n`)
-          }
-        } else {
-          ctx.log.debug('SSE: sending stdout data')
-          res.write(`event: stdout\ndata: ${line}\n\n`)
-        }
-      }
-    }
-  }
-
-  function onStderr (data) {
-    if (closed) return
-    for (let line of data.toString().split('\n')) {
-      ctx.log.debug('SSE: sending stderr data')
-      res.write(`event: stderr\ndata: ${line}\n\n`)
-    }
-  }
-
-  function onExit (data) {
-    if (closed) return
-    ctx.log.debug('SSE: sending end event')
-    res.write(`event: end\ndata: ${data}\n\n`)
-  }
-}
 
 // Proxy/redirect requests to a container session
 function proxyToSession (req, res, ctx) {
